+++ src/routes/tests/field-number/+page.server.ts
@@ -1,3 +1,7 @@
+/**
+ * Fixed: https://github.com/ciscoheat/sveltekit-superforms/issues/85
+ */
+
 import type { PageServerLoad } from './$types';
 import { superValidate } from '$lib/server';
 import { schemaDto } from './schema';