+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.7] - 2023-04-03
 
 ### Fixed
 
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Implicit default value for `ZodRecord` (empty object)
 - UTC date/time formats for `dateProxy`.
+- Form-level Zod issues, can be added with `refine` on the schema, accessed by `$errors._errors`.
 
 ## [0.6.6] - 2023-03-31